tests working for post and get request; next steps are delete and put request tests
